refactor: moved db files to internal dir
