+++ internal/repository/sqlc/migrate.go
@@ -1,4 +1,4 @@
-package db
+package sqlc
 
 import (
 	"github.com/golang-migrate/migrate/v4"
