+++ internal/repository/sqlc/querier.go
@@ -2,7 +2,7 @@
 // versions:
 //   sqlc v1.25.0
 
-package db
+package sqlc
 
 import (
 	"context"
