+++ internal/repository/sqlc/store.go
@@ -1,4 +1,4 @@
-package db
+package sqlc
 
 import (
 	"github.com/jackc/pgx/v5/pgxpool"
