+++ internal/repository/sqlc/user_test.go
@@ -1,4 +1,4 @@
-package db
+package sqlc
 
 import (
 	"context"
