+++ sqlc.yaml
@@ -5,8 +5,8 @@ sql:
   engine: "postgresql"
   gen:
     go: 
-      package: "db"
-      out: "db/sqlc"
+      package: "sqlc"
+      out: "internal/repository/sqlc"
       sql_package: "pgx/v5"
       emit_json_tags: true
       emit_interface: true
